requirements.txt
flask
ultralytics
pillow 